+++ templates/partials/pilot/pilot-information.php
@@ -23,6 +23,6 @@
         <?php echo $data['characterName']; ?>
     </span>
     <span class="eve-intel-pilot-links-wrapper">
-        <small><a class="eve-intel-information-link" href="https://evewho.com/pilot/<?php echo \urlencode($data['characterName']); ?>" target="_blank">evewho <i class="fa fa-external-link" aria-hidden="true"></i></a> | <a class="eve-intel-information-link" href="https://zkillboard.com/character/<?php echo $data['characterID']; ?>/" target="_blank">zkillboard <i class="fa fa-external-link" aria-hidden="true"></i></a></small>
+        <small><a class="eve-intel-information-link" href="https://evewho.com/pilot/<?php echo \urlencode($data['characterName']); ?>" target="_blank" rel="noopener noreferer">evewho <i class="fa fa-external-link" aria-hidden="true"></i></a> | <a class="eve-intel-information-link" href="https://zkillboard.com/character/<?php echo $data['characterID']; ?>/" target="_blank" rel="noopener noreferer">zkillboard <i class="fa fa-external-link" aria-hidden="true"></i></a></small>
     </span>
 </span>